Update DAPNET2APRSNotifier.py
Minor formatting changes to the aprs message to better delineate the message (since messages get CALLSIGN: prepended to them, it was looking like a malformed APRS packet, now we prepent MSG and enclose the forwarded message in brackets, so it can't possibly result in a malformed APRS packet.   In addition added a handler to automatically split messages that would exceed the APRS 67 character limit.